feature: Adds tRPC architecture recfactor to have only one tRPC instance
