+++ src/server/trpc/router.ts
@@ -1,33 +1,6 @@
-import { initTRPC, TRPCError } from "@trpc/server";
-import { Context } from "./context";
-import superjson from "superjson";
 import { z } from "zod";
+import { router, publicProcedure } from "./trpc";
 import { portfolioRouter } from "./routers/portfolio";
-
-const t = initTRPC.context<Context>().create({
-  transformer: superjson, // Configure superjson for data transformation
-  errorFormatter({ shape }) {
-    return shape;
-  },
-});
-
-// Base router and procedure helpers
-export const router = t.router;
-export const publicProcedure = t.procedure;
-
-// Protected procedure helper that requires authentication
-export const protectedProcedure = t.procedure.use(({ ctx, next }) => {
-  if (!ctx.session || !ctx.session.user) {
-    throw new TRPCError({ code: "UNAUTHORIZED" });
-  }
-  return next({
-    ctx: {
-      ...ctx,
-      // infers the `session` as non-nullable
-      session: { ...ctx.session, user: ctx.session.user },
-    },
-  });
-});
 
 export const appRouter = router({
   hello: publicProcedure
